Minor refactoring to reducers and dependant components
