+++ index.js
@@ -5,13 +5,13 @@ import { Provider } from 'react-redux'
 import createSagaMiddleware from 'redux-saga'
 
 import requestReducer from './store/reducers/request'
-import verifyReducer from './store/reducers/verify'
+import initReducer from './store/reducers/init'
 import { watchAuth } from './store/sagas'
 import App from './components/App'
 import { name } from './app.json'
 
 const sagaMiddleware = createSagaMiddleware()
-const store = createStore(combineReducers({ requestReducer, verifyReducer }),
+const store = createStore(combineReducers({ requestReducer, initReducer }),
   applyMiddleware(sagaMiddleware))
 sagaMiddleware.run(watchAuth)
 const AppContainer = () => (
