+++ store/actions/index.js
@@ -1,3 +1,3 @@
 export { authSignIn, authVerify } from './auth'
 export { verifyStart, verifyEnd } from './verify'
-export { requestStart, requestFail, requestSuccess, clearRequestError } from './request'
+export { requestStart, requestFail, requestSuccess } from './request'
