+++ store/reducers/request.js
@@ -1,4 +1,4 @@
-import { REQUEST_START, REQUEST_SUCCESS, REQUEST_FAIL, CLEAR_REQUEST_ERROR } from '../actions/action-types'
+import { REQUEST_START, REQUEST_SUCCESS, REQUEST_FAIL } from '../actions/action-types'
 
 const initState = {
   error: null,
@@ -14,8 +14,6 @@ export default (state = initState, { type, ...payload }) => {
     case REQUEST_FAIL:
       const { error } = payload
       return { ...state, loading: false, error }
-    case CLEAR_REQUEST_ERROR:
-      return { ...state, error: null }
     default:
       return state
   }
